% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cod_municipios_ine}
\alias{cod_municipios_ine}
\title{Diccionario municipios del INE}
\format{
Un data frame con 8.121 filas y 7 columnas.
\describe{
}
}
\source{
INE
}
\usage{
cod_municipios_ine
}
\description{
Diccionario municipios del INE
}
\examples{
data(cod_municipios_ine)
}
\keyword{datasets}
