% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_bomberos_long}
\alias{data_bomberos_long}
\title{Datos con intervenciones de bomberos en la ciudad de Madrid}
\format{
Un data frame con 1.848 filas y 6 columnas.
\describe{
}
}
\source{
Ayuntamiento de Madrid
}
\usage{
data_bomberos_long
}
\description{
Datos con intervenciones de bomberos en la ciudad de Madrid
}
\examples{
data(data_bomberos_long)
}
\keyword{datasets}
