% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_bomberos_monthly}
\alias{data_bomberos_monthly}
\title{Datos con intervenciones de bomberos en la ciudad de Madrid (mensuales)}
\format{
Un data frame con 83 filas y 3 columnas.
\describe{
}
}
\source{
Ayuntamiento de Madrid
}
\usage{
data_bomberos_monthly
}
\description{
Datos con intervenciones de bomberos en la ciudad de Madrid (mensuales)
}
\examples{
data(data_bomberos_monthly)
}
\keyword{datasets}
