% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diputados_congreso}
\alias{diputados_congreso}
\title{Resultados del PP en las elecciones de 2020 en municipios de menos de 20.000 habitantes}
\format{
Un data frame con 350 filas y 9 columnas.
\describe{
}
}
\source{
Congreso de los Diputados
}
\usage{
diputados_congreso
}
\description{
Resultados del PP en las elecciones de 2020 en municipios de menos de 20.000 habitantes
}
\examples{
data(diputados_congreso)
}
\keyword{datasets}
