% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edades}
\alias{edades}
\title{Datos de edad media por nombre en España}
\format{
Un data frame con 54.364 filas y 5 columnas.
\describe{
}
}
\source{
INE
}
\usage{
edades
}
\description{
Datos de edad media por nombre en España
}
\examples{
data(edades)
}
\keyword{datasets}
