% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{galicia_20_munis}
\alias{galicia_20_munis}
\title{Resultados por municipio de las elecciones de 2020 en Galicia}
\format{
Un data frame con 1.868 filas y 13 columnas.
\describe{
}
}
\source{
Instituto Galego de Estadística
}
\usage{
galicia_20_munis
}
\description{
Resultados por municipio de las elecciones de 2020 en Galicia
}
\examples{
data(galicia_20_munis)
}
\keyword{datasets}
