% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{galicia_20_pp_menos_20k}
\alias{galicia_20_pp_menos_20k}
\title{Resultados del PP en las elecciones de 2020 en municipios de menos de 20.000 habitantes}
\format{
Un data frame con 296 filas y 13 columnas.
\describe{
}
}
\source{
Instituto Galego de Estadística
}
\usage{
galicia_20_pp_menos_20k
}
\description{
Resultados del PP en las elecciones de 2020 en municipios de menos de 20.000 habitantes
}
\examples{
data(galicia_20_pp_menos_20k)
}
\keyword{datasets}
