% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{galicia_24_tam_censo}
\alias{galicia_24_tam_censo}
\title{Resultados de cada partido en cada tipo de municipio}
\format{
Un data frame con 19 filas y 8 columnas.
\describe{
}
}
\source{
Wikipedia
}
\usage{
galicia_24_tam_censo
}
\description{
Resultados de cada partido en cada tipo de municipio
}
\examples{
data(galicia_24_tam_censo)
}
\keyword{datasets}
