% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{galicia_psoe_bng_long}
\alias{galicia_psoe_bng_long}
\title{Resultados del BNG y el PSOE en formato long}
\format{
Un data frame con 24 filas y 2 columnas.
\describe{
}
}
\source{
Xunta de Galicia
}
\usage{
galicia_psoe_bng_long
}
\description{
Resultados del BNG y el PSOE en formato long
}
\examples{
data(galicia_psoe_bng_long)
}
\keyword{datasets}
