% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{historic_totales_galicia}
\alias{historic_totales_galicia}
\title{Resultados totales de las elecciones gallegas desde 1981}
\format{
Un data frame con 57 filas y 3 columnas.
\describe{
}
}
\source{
Instituto Galego de Estadística
}
\usage{
historic_totales_galicia
}
\description{
Resultados totales de las elecciones gallegas desde 1981
}
\examples{
data(historic_totales_galicia)
}
\keyword{datasets}
