% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtcars_to_plot}
\alias{mtcars_to_plot}
\title{Estadísticas de 9 modelos de coche diferentes}
\format{
Un data frame con 99 filas y 4 columnas.
\describe{
}
}
\source{
Elaboración propia
}
\usage{
mtcars_to_plot
}
\description{
Estadísticas de 9 modelos de coche diferentes
}
\examples{
data(mtcars_to_plot)
}
\keyword{datasets}
