% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{renta_munis}
\alias{renta_munis}
\title{Renta media por persona en todos los municipios de España}
\format{
Un data frame con 8.136 filas y 5 columnas.
\describe{
}
}
\source{
Instituto Nacional de Estadística
}
\usage{
renta_munis
}
\description{
Renta media por persona en todos los municipios de España
}
\examples{
data(renta_munis)
}
\keyword{datasets}
