% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contaminacion_madrid}
\alias{contaminacion_madrid}
\title{Datos horarios de contaminación en la ciudad de Madrid}
\format{
Un data frame con 4.005 filas y 56 columnas.
\describe{
}
}
\source{
Ayuntamiento de Madrid
}
\usage{
contaminacion_madrid
}
\description{
Datos horarios de contaminación en la ciudad de Madrid
}
\examples{
data(contaminacion_madrid)
}
\keyword{datasets}
