% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_ccaa}
\alias{covid_ccaa}
\title{Contagios acumulados e incidencia a 14 días por 100.000 habitantes en cada comunidad autónoma}
\format{
Un data frame con 14.308 filas y 4 columnas.
\describe{
}
}
\source{
Ministerio de Sanidad y EL PAÍS
}
\usage{
covid_ccaa
}
\description{
Contagios acumulados e incidencia a 14 días por 100.000 habitantes en cada comunidad autónoma
}
\examples{
data(covid_ccaa)
}
\keyword{datasets}
