% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_madrid}
\alias{covid_madrid}
\title{Contagios acumulados e incidencia a 14 días por 100.000 habitantes de la Comunidad de Madrid}
\format{
Un data frame con 753 filas y 4 columnas.
\describe{
}
}
\source{
Ministerio de Sanidad y EL PAÍS
}
\usage{
covid_madrid
}
\description{
Contagios acumulados e incidencia a 14 días por 100.000 habitantes de la Comunidad de Madrid
}
\examples{
data(covid_madrid)
}
\keyword{datasets}
