% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_bomberos_all}
\alias{data_bomberos_all}
\title{Datos con intervenciones de bomberos en la ciudad de Madrid (total)}
\format{
Un data frame con 7 filas y 3 columnas.
\describe{
}
}
\source{
Ayuntamiento de Madrid
}
\usage{
data_bomberos_all
}
\description{
Datos con intervenciones de bomberos en la ciudad de Madrid (total)
}
\examples{
data(data_bomberos_all)
}
\keyword{datasets}
