% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{galicia_24_munis}
\alias{galicia_24_munis}
\title{Resultados de cada partido en cada municipio en Galicia}
\format{
Un data frame con 3212 filas y 11 columnas.
\describe{
}
}
\source{
Wikipedia
}
\usage{
galicia_24_munis
}
\description{
Resultados de cada partido en cada municipio en Galicia
}
\examples{
data(galicia_24_munis)
}
\keyword{datasets}
