% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{galicia_evol}
\alias{galicia_evol}
\title{Resultados totales en las elecciones de Galicia en todas las legislaturas}
\format{
Un data frame con 52 filas y 3 columnas.
\describe{
}
}
\source{
Wikipedia
}
\usage{
galicia_evol
}
\description{
Resultados totales en las elecciones de Galicia en todas las legislaturas
}
\examples{
data(galicia_evol)
}
\keyword{datasets}
