% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{galicia_psoe_bng_wide}
\alias{galicia_psoe_bng_wide}
\title{Resultados del BNG y el PSOE en formato wide}
\format{
Un data frame con 12 filas y 2 columnas.
\describe{
}
}
\source{
Xunta de Galicia
}
\usage{
galicia_psoe_bng_wide
}
\description{
Resultados del BNG y el PSOE en formato wide
}
\examples{
data(galicia_psoe_bng_wide)
}
\keyword{datasets}
