% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{galicia_total_2024}
\alias{galicia_total_2024}
\title{Resultados totales en las elecciones de Galicia de 2024}
\format{
Un data frame con 6 filas y 3 columnas.
\describe{
}
}
\source{
Xunta de Galicia
}
\usage{
galicia_total_2024
}
\description{
Resultados totales en las elecciones de Galicia de 2024
}
\examples{
data(galicia_total_2024)
}
\keyword{datasets}
