% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gasto_renta_munis}
\alias{gasto_renta_munis}
\title{Gasto por habitante y renta del municipio en municipios de más de 20.000 habitantes}
\format{
Un data frame con 409 filas y 6 columnas.
\describe{
}
}
\source{
Elaboración propia
}
\usage{
gasto_renta_munis
}
\description{
Gasto por habitante y renta del municipio en municipios de más de 20.000 habitantes
}
\examples{
data(gasto_renta_munis)
}
\keyword{datasets}
