% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{historic_totales_galicia}
\alias{historic_totales_galicia}
\title{Resultados totales de las elecciones gallegas desde 1981}
\format{
Un data frame con 57 filas y 3 columnas.
\describe{
\item{election}{Año de la elección}
\item{partido}{Partido político}
\item{votos}{Número de votos obtenidos}
}
}
\source{
Instituto Galego de Estadística
}
\usage{
historic_totales_galicia
}
\description{
Serie histórica de resultados electorales en las elecciones autonómicas de Galicia
desde 1981 hasta la actualidad.
}
\examples{
data(historic_totales_galicia)
}
\keyword{datasets}
