% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{renta_munis_madrid}
\alias{renta_munis_madrid}
\title{Renta media por persona en los municipios de la Comunidad de Madrid}
\format{
Un data frame con 179 filas y 5 columnas.
\describe{
}
}
\source{
Instituto Nacional de Estadística
}
\usage{
renta_munis_madrid
}
\description{
Renta media por persona en los municipios de la Comunidad de Madrid
}
\examples{
data(renta_munis_madrid)
}
\keyword{datasets}
